remove irc mention from TSC meeting template (#106)
* remove irc mention from TSC meeting template

I don't have statistics to back it up, but I don't believe the statement that "many of us will be available on IRC during the meeting" is still true. We also don't typically bring up questions from IRC during meetings, so the statement is inaccurate and could lead the live meeting audience to ask questions in the wrong place. Therefore removing it seems like a good idea.