Update invited_tsc
Yagiz is the champion for the performance strategic initiative. They should be invited to the meetings to give updates or surface anything that is blocking their work. There is no requirement to attend and it would be totally fine for Yagiz to leave updates in the meeting issues as appropriate and never attend at all if they so choose. But they should also have the option to attend if it is useful to them.